+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2025-12-11
+
+### Added
+
+- TestKit properties support via `testKitPropertyName` parameter in `buildGradleRunner` function
+- Integration with [CoverJet](https://github.com/gw-kit/cover-jet-plugin) plugin through `io.github.gwkit.coverjet.test-kit` system property
+- Automatic injection of TestKit properties into test project's `gradle.properties`
+
 ## [0.0.1] - 2025-12-06
 
 ### Added